Clean up Rest class by extracting DelayInterceptor
